message board functional
