+++ bc-react/src/actions/MessageActions.js
@@ -1,59 +1,55 @@
 import dispatcher from '../Dispatcher'
 
-let baseUrl;
+let apiUrl;
 if(process.env.NODE_ENV === 'production'){
-  baseUrl = "/"
+  apiUrl = "/"
 } else {
-  baseUrl = "http://localhost:4000/"
+  apiUrl = "http://localhost:4000/"
 }
 
-export function updateMessageInput(attribute, value){
-  dispatcher.dispatch({
-    type: 'UPDATE_MESSAGE_INPUT',
-    attribute: attribute,
-    value: value
-  })
-}
-
-export function sendMessage(messageAttributes){
-  dispatcher.dispatch({
-    type: 'SEND_MESSAGE'
-  })
-}
-
-export function processMessage(attributes){
+export function addMessage(attributes){
   const params = {
-    //this is the action we send over
     method: 'POST',
     headers: {'Content-Type': 'application/json'},
     body: JSON.stringify(attributes)
   }
-  //fetch is a new feature of ES6, it's a standard for fetching data asyncrhonously using promises
-  fetch(`${baseUrl}messages`, params).then((response)=>{
-    if(response.ok){0
-      response.json().then((body)=>{
-        //this is the message we get back in postman
-        //here we handle it back on the server
-        //this returns the message that was just created
+  fetch(`${apiUrl}messages`, params).then(function(response){
+    if(response.status === 200){
+      response.json().then(function(body){
+        // send the message to the store
         dispatcher.dispatch({
-          type: 'UPDATE_MESSAGE_DETAIL',
-          attributes: body.message
-        })
-      })
-    }else{
-      response.json().then((body)=>{
-        dispatcher.dispatch({
-          type: 'MESSAGE_FAIL',
-          errors: body.errors
+          type: 'ADD-MESSAGE',
+          message: body.message
         })
       })
     }
+  }).catch(function(error){
+    console.log("There was an error: " + error);
   })
 }
 
-// export function updateMessages(attributes){
-//   dispatcher.dispatch({
-//     type: 'UPDATE_MESSAGES',
-//     attributes: attributes
-//   })
-// }
+
+export function fetchMessages(){
+  let success;
+  const params = {
+        method: 'GET',
+        headers: {'Content-Type': 'application/json'}
+      }
+  fetch(`${apiUrl}messages`, params)
+    .then((response)=>{
+      success = response.ok
+      return response.json()
+    })
+    .then((body)=>{
+      if (success){
+        let messages = body.messages
+        dispatcher.dispatch({
+          type: "FETCH-MESSAGES",
+          messages: messages
+        })
+      }
+      else {
+        console.log("failure!", body)
+      }
+    })
+}
