+++ bc-react/src/components/MessageBoardToggle.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import MessageBoardInput from './MessageBoardInput';
+import MessageBoard from './MessageBoard';
 import ToggleDisplay from 'react-toggle-display';
 
 class MessageBoardToggle extends Component {
@@ -15,7 +15,7 @@ class MessageBoardToggle extends Component {
         <div>
           <button className="sticky" type='button' onClick={ () => this.handleToggle() }><img src='../Images/chat.png' title='chat' alt='chat'/></button>
           <ToggleDisplay show={this.state.show}>
-            <MessageBoardInput />
+            <MessageBoard />
           </ToggleDisplay>
         </div>
     );
